// src/app/models/evt-models.ts
export class GenericElement {
    // tslint:disable-next-line: no-any
    type: Type<any>;
    path?: string;
    class?: string;
    attributes: Attributes;
    content: Array<ParseResult<GenericElement>>;
}

export type XMLElement = HTMLElement;
export type OriginalEncodingNodeType = XMLElement;

export interface EditionStructure {
    pages: Page[];
}

export type ViewModeId = 'readingText' | 'imageText' | 'textText' | 'collation' | 'textSources' | 'textVersions';

export interface ViewMode {
    id: ViewModeId;
    icon: string;
    iconSet?: 'evt' | 'far' | 'fas';
    label: string;
    disabled?: boolean;
}

export interface Page {
    id: string;
    label: string;
    facs: string;
    originalContent: OriginalEncodingNodeType[];
    parsedContent: Array<ParseResult<GenericElement>>;
    url: string;
    facsUrl: string;
}
// ...
export interface Attributes { [key: string]: string; }
// ...
export type LemmatizedEntityType = 'lem' | 'w' | 'list' | 'item' | 'entry' ;
export class LemmatizedEntitiesList extends GenericElement {
    id: string;
    label: string;
    lemmatizedEntityType: LemmatizedEntityType;
    description?: Description;
    sublists: LemmatizedEntitiesList[];
    content: LemmatizedEntity[];
    relations: Relation[];
    originalEncoding: OriginalEncodingNodeType;
}

export class LemmatizedEntity extends GenericElement {
    id: string;
    sortKey: string;
    label: LemmatizedEntityLabel;
    lemmatizedEntityType: LemmatizedEntityType | 'interpGrp';
    content: LemmatizedEntityInfo[];
    originalEncoding: OriginalEncodingNodeType;
}

export type LemmatizedEntityLabel = string;

export class LemmatizedEntityInfo extends GenericElement {
    label: string;
}

export interface LemmatizedEntityOccurrence {
    pageId: string;
    pageLabel: string;
    refsByDoc: LemmatizedEntityOccurrenceRef[];
}
export interface LemmatizedEntityOccurrenceRef {
    docId: string;
    docLabel: string;
    refs: GenericElement[];
}
// ...
export class Relation extends GenericElement {
    name?: string;
    activeParts: string[]; // Pointers to entities involved in relation
    mutualParts: string[]; // Pointers to entities involved in relation
    passiveParts: string[]; // Pointers to entities involved in relation
    description: Description;
    relationType?: string;
}

export type Description = Array<ParseResult<GenericElement>>;